third_party/WebKit/LayoutTests/fast/css/cssom-setproperty-undefined-value.html

<!doctype html>
<html>
<head>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
<script>
test(function() {

    var style = document.body.style;
    style.color = 'white';

    assert_equals(style.color, 'white');
    style.setProperty('color', undefined);
    assert_equals(style.color, 'white');
}, "Verify that setting a CSS property to undefined has no effect.");
</script>
</body>
</html>
